resolved input Email component
