+++ src/client/stories/index.js
@@ -27,7 +27,7 @@ import { suggestedRoutesList } from "./suggestedRoutes/suggestedRoutesList";
 import { SuggestedRouteTableStory } from "./SuggestedRouteTableStory";
 
 import { NavBarStory } from "./NavBarStory";
-import { InputTextStory } from "./InputText";
+import { InputEmailStory } from "./InputEmail";
 
 const MapStories = storiesOf("Map", module)
   .add("Map Voyages List", MapVoyagesList)
@@ -46,7 +46,8 @@ const FormStories = storiesOf("Forms elements", module)
   .add("DateTimeComponent", DateTimeStory)
   .add("InputFieldNumberComponent", InputFieldNumberStory)
   .add("DropdownComponent", DropdownStory)
-  .add("CheckboxComponent", CheckboxStory);
+  .add("CheckboxComponent", CheckboxStory)
+  .add("InputEmail", InputEmailStory);
 FormStories.addDecorator(withKnobs);
 
 const VoyageStories = storiesOf("Start Voyage", module).add(
@@ -87,9 +88,3 @@ storiesOf("Header", module).add("Navigation Bar", NavBarStory);
 storiesOf("Voyage Summary", module)
   .add("All Voyage Summary Tables", VoyagesSummaryStory)
   .addDecorator(withKnobs);
-
-const InputTextStories = storiesOf("Input Text", module).add(
-  "InputText",
-  InputTextStory
-);
-InputTextStories.addDecorator(withKnobs);
